Set the UI to use a newly refreshed token if it is provided
